Fix issue with neovim
Vim was being super annoying because when you hit `ESC` followed by `j`
(or `k`) to exit edit mode and move up/down it would move switch the
lines instead of moving the cursor. If you hit `ESC` and wait a little
bit it would work correctly. The problem was this issue with tmux.
Luckily Gemini found the answer pretty fast.